# CCIP-2207 Boosting LogPoller performance by fixing how order by clauses are written