--- a/.changeset/proud-toys-travel.md
+++ b/.changeset/proud-toys-travel.md
@@ -1,0 +1,5 @@
+---
+"chainlink": patch
+---
+
+Improving LogPoller read queries by properly sorting by multiple columns #updated