@0xnogo 2024-04-29
Very nice improvement! 
<nit> I'm curious if you have tried running the `EXPLAIN ANALYZE` command on the query both with and without the parentheses. The performance results you've provided already give a good indication of how these changes have helped. `EXPLAIN ANALYZE` can offer additional insights into what happens during the query execution (inefficient joins, if indexes are being used...).