@reductionista 2024-04-30
@0xnogo 
> Not in the scope of this PR, but I am seeing some SELECT *. It can also potentially be an offender of runtime performance. We could potentially check if we require all the columns.

We do need all columns. The api methods for LogPoller have to be generic enough to support any possible chainlink product's needs.  So they return one data structure representing a Log and one representing a Block.  The purpose of the logs and log_poller_blocks table is to allow products to search for a set of logs or blocks based on specific criteria.  These data structs are not large, but LogPoller must return them each as a whole not individual fields (columns). These will soon be in entirely different repos on different release schedules, and run as separate executables so LogPoller must stay as general as possible and not make any assumptions about which fields the products may or may not need to use for a specific application. The filters table is to persist filters across node restarts--the only time it reads from that table is to load the entire table back into memory as the node is starting up.